+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added [Graphics Tier](https://docs.unity3d.com/ScriptReference/Rendering.GraphicsTier.html) information to reported Shader Variants
 * Added diagnostic message formatting support
 * Added dependencies panel to assembly view
+* Added *ImporterType* to Build File properties
 * Fixed reporting of assemblies not compiled due to dependencies
 * Improved code diagnostic messages